+++ .gitignore
@@ -30,3 +30,4 @@ wled-update.sh
 /wled00/wled00.ino.cpp
 /wled00/html_*.h
 /wled00/js_*.h
+/wled00/idf_component.yml
