+++ platformio.ini
@@ -830,7 +830,7 @@ lib_ignore =
   Matter
 custom_usermods =
   ${common.default_usermods}
-  usermod_v2_matter
+  matter_over_wifi
 board_build.partitions = tools/WLED_ESP32_8MB_3MB_APP.csv
 board_build.flash_mode = qio
 board_upload.flash_size = 8MB
