+++ usermods/matter_over_wifi/library.json
@@ -1,5 +1,5 @@
 {
-  "name": "usermod_v2_matter",
+  "name": "matter_over_wifi",
   "build": {
     "libArchive": false,
     "flags": ["-I${platformio.src_dir}"]
